feat(endpoint): Make global endpoint for finding all the things
